MS-345-346: acosh/sqrt/log/exp/erf/var_axis bwd parity + bench 160
## activation/mod.rs: doc comments for basic/parametric submodules

## Parity tests: PyTorch 292 (+6), JAX 145 (+3)
PyTorch:
- test_acosh_bwd: d/dx acosh(x) = 1/sqrt(x^2-1)
- test_sqrt_bwd: d/dx sqrt(x) = 1/(2*sqrt(x))
- test_log_bwd: d/dx ln(x) = 1/x
- test_exp_bwd: d/dx e^x = e^x
- test_erf_bwd: d/dx erf(x) = (2/sqrt(pi))*exp(-x^2)
- test_var_axis_bwd: var(dim=1, unbiased=True) fwd+bwd

JAX: test_acosh_bwd / test_sqrt_bwd / test_erf_bwd

## G-043 bench: 152 -> 160 rows
sin_bwd, cos_bwd, tan_bwd, sqrt_bwd, log_bwd,
exp_bwd, softmax_bwd, var_bwd (all fwd+bwd benchmarks)

255/255 coeus-ops+autograd tests passing.
